LPS-165757 Avoid Overflow exception when trying to bind Double.MAX_VALUE in Oracle

+++ modules/apps/headless/headless-admin-address/headless-admin-address-test/src/testIntegration/java/com/liferay/headless/admin/address/resource/v1_0/test/RegionResourceTest.java
@@ -332,7 +332,7 @@ public class RegionResourceTest extends BaseRegionResourceTestCase {
 	protected Region testGraphQLRegion_addRegion() throws Exception {
 		Region region = randomRegion();
 
-		region.setPosition(Double.MAX_VALUE);
+		region.setPosition((double)Integer.MAX_VALUE);
 
 		return _addRegion(region);
 	}
